fix(keyboard-shortcuts): Rename .web/.native files.
The keyboard shortcuts feature is used only on web. We don't need the suffixes.
